pyproject.toml
[tool.poetry]
name = "domainsale"
version = "0.1.0"
description = "A secure library for discovering and displaying 'for-sale' status of internet domains"
authors = ["DomainSaleGPT <info@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/example/domainsale"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: Security",
]
packages = [{include = "domainsale"}]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
dnspython = ">=2.0.0"
requests = ">=2.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
flake8 = "^6.0.0"
pylint = "^2.17.0"
black = "^23.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
domainsale = "domainsale.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["domainsale/tests"]
python_files = "test_*.py"
